fixed rbt tree and formats of numbers changed to long long

+++ hash_chaining.c
@@ -5,19 +5,19 @@
 
 
 
-HashTable* create_table(int size){
+HashTable* create_table(long long size){
     HashTable *ht = (HashTable*)malloc(sizeof(HashTable));
     ht->size = size;
     ht->count = 0;
     ht->table = (Node**)malloc(size * sizeof(Node*));
-    for(int i = 0; i < size; i++){
+    for(long long i = 0; i < size; i++){
         ht->table[i] = NULL;
     }
     return ht;
 }
 
-void insert(HashTable *ht, int key){
-    int index = key % ht->size; // jednoduchá hashovací funkce
+void insert(HashTable *ht, long long key){
+    long long index = key % ht->size; // jednoduchá hashovací funkce
     Node *new_node = (Node*)malloc(sizeof(Node));
     new_node->key = key;
     new_node->next = ht->table[index]; // vkládáme na začátek řetězce
@@ -30,8 +30,8 @@ void insert(HashTable *ht, int key){
     
 }
 
-Node* search(HashTable *ht, int key){
-    int index = key % ht->size;
+Node* search(HashTable *ht, long long key){
+    long long index = key % ht->size;
     Node *current = ht->table[index];
     while(current != NULL){
         if(current->key == key){
@@ -42,8 +42,8 @@ Node* search(HashTable *ht, int key){
     return NULL; // klíč nenalezen
 }
 
-void delete_key(HashTable *ht, int key){
-    int index = key % ht->size;
+void delete_key(HashTable *ht, long long key){
+    long long index = key % ht->size;
     Node *current = ht->table[index];
     Node *prev = NULL;
     while(current != NULL){
@@ -66,14 +66,14 @@ void delete_key(HashTable *ht, int key){
 }
 
 void rehash(HashTable *ht, double factor){
-    int old_size = ht->size;
+    long long old_size = ht->size;
     Node **old_table = ht->table;
 
-    int new_size = (int)(old_size * factor);
+    long long new_size = (int)(old_size * factor);
     HashTable *new_ht = create_table(new_size);
 
     // Prenesieme všetky kľúče do novej tabuľky
-    for (int i = 0; i < old_size; i++) {
+    for (long long i = 0; i < old_size; i++) {
         Node *current = old_table[i];
         while (current != NULL) {
             insert(new_ht, current->key);
@@ -81,7 +81,7 @@ void rehash(HashTable *ht, double factor){
         }
     }
     // Uvoľníme uzly v starej tabuľke
-    for (int i = 0; i < old_size; i++) {
+    for (long long i = 0; i < old_size; i++) {
         Node *current = old_table[i];
         while (current != NULL) {
             Node *tmp = current;
@@ -103,7 +103,7 @@ void rehash(HashTable *ht, double factor){
 void free_table(HashTable *ht) {
     if (ht == NULL) return;
 
-    for (int i = 0; i < ht->size; i++) {
+    for (long long i = 0; i < ht->size; i++) {
         Node *current = ht->table[i];
 
         while (current != NULL) {
